+++ src/backend/shared/utils/common.c
@@ -3,7 +3,7 @@
 #include <ctype.h>
 
 char *strip(const char *s) {
-  size_t size = strlen(s);
+  int size = strlen(s);
   var copy = str(s);
   if (!size) return copy;
 
@@ -21,7 +21,7 @@ char *str(const char *format, ...) {
   va_list arguments, size_check;
   va_start(arguments, format);
   va_copy(size_check, arguments);
-  size_t size = vsnprintf(NULL, 0, format, size_check);
+  int size = vsnprintf(NULL, 0, format, size_check);
   char *buffer = (char *) malloc(size + 1);
   buffer[0] = '\0';
   vsprintf(buffer, format, arguments);
@@ -31,7 +31,7 @@ char *str(const char *format, ...) {
 }
 
 bool starts_with(const char *prefix, const char *s) {
-  size_t str_length = strlen(s);
-  size_t prefix_length = strlen(prefix);
+  int str_length = strlen(s);
+  int prefix_length = strlen(prefix);
   return str_length < prefix_length ? false : memcmp(prefix, s, prefix_length) == 0;
 }
